更依赖; 修复误用 atof 产生 double 转 float 警告, 改用 std::stof; 增加常用类型定义 PropertyEditorCommon.h 并加入 Precompiled Headers

+++ Source/PropertyEditor/src/ExampleStructure.cpp
@@ -10,7 +10,7 @@ void LoadCVector3FromRwNode(const CRwNode* rw, CVector3& vec)
 {
 	auto str = GetRwValueAs<Niflect::CString>(rw->GetValue());
 	auto vecEntryStr = NiflectUtil::Split(str, ' ');
-	vec.m_x = atof(vecEntryStr[0].c_str());
-	vec.m_y = atof(vecEntryStr[1].c_str());
-	vec.m_z = atof(vecEntryStr[2].c_str());
+	vec.m_x = std::stof(vecEntryStr[0].c_str());
+	vec.m_y = std::stof(vecEntryStr[1].c_str());
+	vec.m_z = std::stof(vecEntryStr[2].c_str());
 }
